refactor: Change message_response -> messages
