+++ messenger/controller.py
@@ -1,7 +1,7 @@
 from flask import Blueprint, request
 
 from messenger.verify.service import verify_service
-from messenger.message_response.router import message_response_router
+from messenger.messages.router import message_response_router
 from messenger.messaging_referrals import service as messaging_referrals_service
 
 messenger_controller = Blueprint("messenger", __name__)
